et-rest-api-server/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
black==24.4.2
cassandra-driver==3.29.1
certifi==2023.11.17
click==8.1.7
distlib==0.3.8
Django==4.2.14
django-filter==24.2
djangorestframework==3.15.2
dnspython==2.6.1
filelock==3.13.1
future==1.0.0
geomet==0.2.1.post1
gunicorn==22.0.0
h11==0.14.0
importlib_metadata==8.2.0
isort==5.13.2
kaleido==0.2.1
macholib==1.16.3
Markdown==3.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pipenv==2023.11.15
platformdirs==4.1.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.3
virtualenv==20.25.0
zipp==3.19.2
